# Make the string buffer valid UTF-8 [discussion]