--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -69,7 +69,7 @@ commands:
     steps:
       - cmake_build
       - run: ctest $CTEST_FLAGS -L acceptance
-      - run: ctest $CTEST_FLAGS -LE acceptance -E checkperf
+      - run: ctest $CTEST_FLAGS -LE acceptance
 
   cmake_test_all:
     steps: