--- a/include/simdjson/document.h
+++ b/include/simdjson/document.h
@@ -1038,6 +1038,7 @@ public:
   really_inline bool on_null_atom() noexcept; ///< @private
   really_inline uint8_t *on_start_string() noexcept; ///< @private
   really_inline bool on_end_string(uint8_t *dst) noexcept; ///< @private
+  bool handle_long_string(uint8_t *dst) noexcept;///< @private
   really_inline bool on_number_s64(int64_t value) noexcept; ///< @private
   really_inline bool on_number_u64(uint64_t value) noexcept; ///< @private
   really_inline bool on_number_double(double value) noexcept; ///< @private