--- a/include/simdjson/parsedjson_iterator.h
+++ b/include/simdjson/parsedjson_iterator.h
@@ -70,8 +70,19 @@ public:
   // print_with_escapes) return value is valid UTF-8, it may contain NULL chars
   // within the string: get_string_length determines the true string length.
   inline const char *get_string() const {
+      uint32_t index =  uint32_t(current_val);
+      uint32_t len = (current_val >> 32) & 0xFFFFFF;
+      if(unlikely(len > 0x7fffff)) {
+          uint64_t string_buf_index = current_val;
+          len <<= 9;
+          uint32_t c1 = doc.string_buf[string_buf_index + 1] - 32;
+          len |= c1 << 4;
+          uint32_t c2 = doc.string_buf[string_buf_index + 2] - 32;
+          len |= c2;
+          index -= len;
+      }
       return reinterpret_cast<const char *>(
-          doc.string_buf.get() + uint32_t(current_val));
+          doc.string_buf.get() + index);
   }
 
   // return the length of the string in bytes
@@ -79,11 +90,10 @@ public:
       uint32_t len = (current_val >> 32) & 0xFFFFFF;
       if(unlikely(len > 0x7fffff)) {
           uint64_t string_buf_index = current_val;
-          len ^= 0x800000;
           len <<= 9;
-          uint32_t c1 = doc.string_buf[string_buf_index - 2] - 32;
+          uint32_t c1 = doc.string_buf[string_buf_index + 1] - 32;
           len |= c1 << 4;
-          uint32_t c2 = doc.string_buf[string_buf_index - 1] - 32;
+          uint32_t c2 = doc.string_buf[string_buf_index + 2] - 32;
           len |= c2;
       }
       // if the slow path can be avoided, then we get the string length without