--- a/src/document_parser_callbacks.h
+++ b/src/document_parser_callbacks.h
@@ -87,60 +87,55 @@ really_inline bool parser::on_null_atom() noexcept {
 }
 
 really_inline uint8_t *parser::on_start_string() noexcept {
-  /* We advance the pointer. */
-  // If we limit JSON documents to strictly less 4GB of
-  // string content, then current_string_buf_loc
-  // - doc.string_buf.get() fits in 32 bits. This leaves us
-  // three free bytes.
-  uint32_t position = uint32_t(current_string_buf_loc - doc.string_buf.get()); // cannot overflow because documents are limited to < 4GB
-  write_tape(uint64_t(position), internal::tape_type::STRING);
+  /* We do nothing, the actual work occurs in on_end_string() */
   return current_string_buf_loc;
 }
 
+bool parser::handle_long_string(uint8_t *dst) noexcept {
+  uint64_t str_length = dst - current_string_buf_loc;
+  // Oh gosh, we have a long string (8MB). We expect that this is
+  // highly uncommon. We want to keep everything else super efficient,
+  // so we will pay a complexity price for this one uncommon case.
+  int offset = 2;
+  uint64_t position = current_string_buf_loc - doc.string_buf.get() + offset;
+  uint64_t lenmark = uint64_t(0x800000 | (str_length >> 9));
+  uint64_t payload =  position |  (lenmark << 32);
+  write_tape(payload, internal::tape_type::STRING);
+  // We are going to make room. This copy is not free. However,
+  // it allows us to handle the common case with ease and with
+  // relatively little complexity. And a memcopy is not that slow:
+  // it may run at tens of GB/s.
+  // And we expect that it will effectively never happen in practice
+  // so there is no cause to complexify the rest of the code.
+  memmove(current_string_buf_loc + offset, current_string_buf_loc, str_length);
+  dst += offset;
+  // We have three free bytes, but
+  // we need a leading 1, so that's 24-1 = 23. 32-23=9 remaining bits.
+  // We have 9 bits left to code, which we do on the string buffer
+  // using two bytes. We encoding the binary data using ASCII characters.
+  // See https://lemire.me/blog/2020/05/02/encoding-binary-in-ascii-very-fast/
+  // for a more general approach.
+  //
+  // These two bytes will appear right before where the string is.
+  current_string_buf_loc[0] = uint8_t(32 + ((str_length & 0x1f0) >> 4)); // (0x1f0>>4)+32 = 63
+  current_string_buf_loc[1] = uint8_t(32 + (str_length & 0xf)); // 32 + 0xf = 47
+  *dst = 0;
+  current_string_buf_loc = dst + 1;
+  return true;
+
+}
+
 really_inline bool parser::on_end_string(uint8_t *dst) noexcept {
-  uint32_t str_length = uint32_t(dst - current_string_buf_loc);
+  uint64_t str_length = dst - current_string_buf_loc;
+  if(unlikely(str_length > 0x7fffff)) {
+    return handle_long_string(dst);
+  }
+  uint64_t position = current_string_buf_loc - doc.string_buf.get();
   // Long document support: Currently, simdjson supports only document
   // up to 4GB.
   // Should we change this constraint, we should then check for overflow in case
   // someone has a crazy string (>=4GB?).
-
-  // We have two scenarios here. Either the string length is
-  // less than 0x7fffff in which case, we have room in the string
-  // header and all is good. Otherwise, we can encode the
-  // string length in the document itself, taking care to
-  // ensure that we do so in ASCII.
-  if(likely(str_length <= 0x7fffff)) { // likely
-    doc.tape[current_loc-1] |=  uint64_t(str_length) << 32;
-  } else {
-    // Oh gosh, we have a long string (8MB). We expect that this is
-    // highly uncommon. We want to keep everything else super efficient,
-    // so we will pay a complexity price for this one uncommon case.
-    int offset = 2;
-    uint64_t payload = doc.tape[current_loc-1] & 0xFFFFFFFF;
-    payload += offset;
-    payload |=  uint64_t(0x800000 | (str_length >> 9)) << 32;
-    doc.tape[current_loc-1] =  payload  | ((uint64_t(char(internal::tape_type::STRING))) << 56);
-    // We are going to make room. This copy is not free. However,
-    // it allows us to handle the common case with ease and with
-    // relatively little complexity. And a memcopy is not that slow:
-    // it may run at tens of GB/s.
-    // And we expect that it will effectively never happen in practice
-    // so there is no cause to complexify the rest of the code.
-    memmove(current_string_buf_loc + offset, current_string_buf_loc, str_length);
-    dst += offset;
-    // We have three free bytes, but
-    // we need a leading 1, so that's 24-1 = 23. 32-23=9 remaining bits.
-    // We have 9 bits left to code, which we do on the string buffer
-    // using two bytes. We encoding the binary data using ASCII characters.
-    // See https://lemire.me/blog/2020/05/02/encoding-binary-in-ascii-very-fast/
-    // for a more general approach.
-    //
-    // These two bytes will appear right before where the string is.
-    current_string_buf_loc[0] = uint8_t(32 + ((str_length & 0x1f0) >> 4)); // (0x1f0>>4)+32 = 63
-    current_string_buf_loc[1] = uint8_t(32 + (str_length & 0xf)); // 32 + 0xf = 47
-  }
-  // NULL termination is still handy if you expect all your strings to
-  // be NULL terminated? It comes at a small cost.
+  write_tape(position | (str_length << 32), internal::tape_type::STRING);
   *dst = 0;
   current_string_buf_loc = dst + 1;
   return true;