Would fix https://github.com/simdjson/simdjson/issues/185

This PR does...

- [X] Move the length of the string on the main tape. This may help performance since you can often determine that you do not have the right key simply by checking the string length. Thus many comparisons can be made without access to the string buffer.
- [X] Make the string buffer a valid string. Currently, we write the string length in the string buffer which makes it non-UTF8 valid. If we only add ASCII content, then we will be able to move the unicode validation off stage 1 and into a cheaper "stage 3". The advantage of that is that the validation would only be done on the actual string content.
- [X] We save a few bytes on the string buffer per string (four bytes). This might help performance a bit (for large files made of strings or at query time).


Negatives...

- [X] We now have to do extra work at query time to deal with the special case of a long string. It is mostly the cost of a never-taken inexpensive branch. It is not a great thing but it is also not disastrous.


Claims of performance regression are overrated. I think we have over-sensitive tests. Regarding performance, let us keep in mind that this PR opens the door to new optimizations, most importantly the removal of UTF8 validation from the main stage 1.



Regarding performance... for each file, the first row is this PR, the second is master...

Trial 1:


```
$ for i in ../jsonexamples/*.json ; do echo $i; ./benchmark/parse -n 100 $i | grep GB | head -n 1; ./../buildmaster/benchmark/parse -n 100 $i | grep GB | head -n 1; done
../jsonexamples/apache_builds.json
|    Speed        :  25.9950 ns per block ( 99.08%) -   0.4062 ns per byte -   4.1818 ns per structural -   2.4616 GB/s
|    Speed        :  25.9472 ns per block ( 98.96%) -   0.4055 ns per byte -   4.1741 ns per structural -   2.4661 GB/s
../jsonexamples/canada.json
|    Speed        :  71.3523 ns per block ( 99.35%) -   1.1149 ns per byte -   7.5056 ns per structural -   0.8969 GB/s
|    Speed        :  71.0723 ns per block ( 99.58%) -   1.1105 ns per byte -   7.4762 ns per structural -   0.9005 GB/s
../jsonexamples/citm_catalog.json
|    Speed        :  22.8718 ns per block ( 99.58%) -   0.3574 ns per byte -   4.5390 ns per structural -   2.7982 GB/s
|    Speed        :  22.8472 ns per block ( 99.44%) -   0.3570 ns per byte -   4.5342 ns per structural -   2.8012 GB/s
../jsonexamples/github_events.json
|    Speed        :  24.4774 ns per block ( 98.32%) -   0.3826 ns per byte -   5.3518 ns per structural -   2.6139 GB/s
|    Speed        :  24.2387 ns per block ( 97.99%) -   0.3788 ns per byte -   5.2996 ns per structural -   2.6396 GB/s
../jsonexamples/google_maps_api_compact_response.json
|    Speed        :  77.8378 ns per block ( 98.87%) -   1.2191 ns per byte -   4.6198 ns per structural -   0.8203 GB/s
|    Speed        :  76.4378 ns per block ( 98.71%) -   1.1972 ns per byte -   4.5367 ns per structural -   0.8353 GB/s
../jsonexamples/google_maps_api_response.json
|    Speed        :  42.3701 ns per block ( 98.46%) -   0.6623 ns per byte -   5.5460 ns per structural -   1.5099 GB/s
|    Speed        :  41.9608 ns per block ( 98.43%) -   0.6559 ns per byte -   5.4925 ns per structural -   1.5246 GB/s
../jsonexamples/gsoc-2018.json
|    Speed        :  19.2063 ns per block ( 98.89%) -   0.3001 ns per byte -  13.1682 ns per structural -   3.3322 GB/s
|    Speed        :  19.5210 ns per block ( 98.89%) -   0.3050 ns per byte -  13.3840 ns per structural -   3.2785 GB/s
../jsonexamples/instruments.json
|    Speed        :  29.4060 ns per block ( 99.43%) -   0.4595 ns per byte -   3.7259 ns per structural -   2.1764 GB/s
|    Speed        :  28.8954 ns per block ( 99.02%) -   0.4515 ns per byte -   3.6612 ns per structural -   2.2148 GB/s
../jsonexamples/marine_ik.json
|    Speed        :  75.3143 ns per block ( 99.45%) -   1.1768 ns per byte -   5.4601 ns per structural -   0.8498 GB/s
|    Speed        :  76.3359 ns per block ( 99.31%) -   1.1928 ns per byte -   5.5342 ns per structural -   0.8384 GB/s
../jsonexamples/mesh.json
|    Speed        :  74.2141 ns per block ( 98.48%) -   1.1597 ns per byte -   5.4748 ns per structural -   0.8623 GB/s
|    Speed        :  76.8083 ns per block ( 98.77%) -   1.2002 ns per byte -   5.6661 ns per structural -   0.8332 GB/s
../jsonexamples/mesh.pretty.json
|    Speed        :  44.6779 ns per block ( 98.76%) -   0.6981 ns per byte -   7.1844 ns per structural -   1.4324 GB/s
|    Speed        :  45.7237 ns per block ( 98.94%) -   0.7145 ns per byte -   7.3525 ns per structural -   1.3997 GB/s
../jsonexamples/numbers100.json
|    Speed        :  67.9624 ns per block ( 99.87%) -   1.0619 ns per byte -   7.9694 ns per structural -   0.9417 GB/s
|    Speed        :  69.6944 ns per block ( 99.17%) -   1.0890 ns per byte -   8.1725 ns per structural -   0.9183 GB/s
../jsonexamples/numbers10.json
|    Speed        :  65.3498 ns per block ( 99.03%) -   1.0211 ns per byte -   7.6633 ns per structural -   0.9793 GB/s
|    Speed        :  67.4611 ns per block ( 99.10%) -   1.0541 ns per byte -   7.9109 ns per structural -   0.9487 GB/s
../jsonexamples/numbers.json
|    Speed        :  65.2907 ns per block ( 98.67%) -   1.0203 ns per byte -   7.6575 ns per structural -   0.9801 GB/s
|    Speed        :  64.6130 ns per block ( 98.40%) -   1.0097 ns per byte -   7.5780 ns per structural -   0.9904 GB/s
../jsonexamples/random.json
|    Speed        :  41.7483 ns per block ( 99.62%) -   0.6524 ns per byte -   3.7837 ns per structural -   1.5328 GB/s
|    Speed        :  40.8523 ns per block ( 99.63%) -   0.6384 ns per byte -   3.7025 ns per structural -   1.5665 GB/s
../jsonexamples/repeat.json
|    Speed        :  49.7921 ns per block ( 98.19%) -   0.7805 ns per byte -   8.7148 ns per structural -   1.2813 GB/s
|    Speed        :  49.0562 ns per block ( 98.19%) -   0.7689 ns per byte -   8.5860 ns per structural -   1.3005 GB/s
../jsonexamples/tree-pretty.json
|    Speed        :  38.8500 ns per block ( 98.57%) -   0.6071 ns per byte -   5.2991 ns per structural -   1.6472 GB/s
|    Speed        :  38.3481 ns per block ( 98.77%) -   0.5993 ns per byte -   5.2306 ns per structural -   1.6687 GB/s
../jsonexamples/twitter10.json
|    Speed        :  29.0067 ns per block ( 99.41%) -   0.4532 ns per byte -   5.1792 ns per structural -   2.2064 GB/s
|    Speed        :  28.9903 ns per block ( 99.58%) -   0.4530 ns per byte -   5.1763 ns per structural -   2.2076 GB/s
../jsonexamples/twitter_api_compact_response.json
|    Speed        :  55.7532 ns per block ( 98.20%) -   0.8743 ns per byte -   7.1735 ns per structural -   1.1437 GB/s
|    Speed        :  55.1646 ns per block ( 98.38%) -   0.8651 ns per byte -   7.0977 ns per structural -   1.1559 GB/s
../jsonexamples/twitter_api_response.json
|    Speed        :  45.2008 ns per block ( 98.69%) -   0.7083 ns per byte -   7.5073 ns per structural -   1.4119 GB/s
|    Speed        :  45.9247 ns per block ( 98.11%) -   0.7196 ns per byte -   7.6275 ns per structural -   1.3897 GB/s
../jsonexamples/twitterescaped.json
|    Speed        :  55.7543 ns per block ( 99.64%) -   0.8712 ns per byte -   8.8661 ns per structural -   1.1478 GB/s
|    Speed        :  56.0678 ns per block ( 99.66%) -   0.8761 ns per byte -   8.9160 ns per structural -   1.1414 GB/s
../jsonexamples/twitter.json
|    Speed        :  26.2190 ns per block ( 99.28%) -   0.4097 ns per byte -   4.6818 ns per structural -   2.4408 GB/s
|    Speed        :  26.8413 ns per block ( 99.19%) -   0.4194 ns per byte -   4.7929 ns per structural -   2.3842 GB/s
../jsonexamples/twitter_timeline.json
|    Speed        :  46.8879 ns per block ( 98.43%) -   0.7327 ns per byte -   5.8093 ns per structural -   1.3647 GB/s
|    Speed        :  45.7470 ns per block ( 98.46%) -   0.7149 ns per byte -   5.6679 ns per structural -   1.3988 GB/s
../jsonexamples/update-center.json
|    Speed        :  33.8374 ns per block ( 99.14%) -   0.5287 ns per byte -   4.4450 ns per structural -   1.8914 GB/s
|    Speed        :  32.9998 ns per block ( 99.46%) -   0.5156 ns per byte -   4.3350 ns per structural -   1.9394 GB/s
```


Trial 2:

```
$ for i in ../jsonexamples/*.json ; do echo $i; ./benchmark/parse -n 100 $i | grep GB | head -n 1; ./../buildmaster/benchmark/parse -n 100 $i | grep GB | head -n 1; done
../jsonexamples/apache_builds.json
|    Speed        :  26.2665 ns per block ( 99.11%) -   0.4105 ns per byte -   4.2255 ns per structural -   2.4362 GB/s
|    Speed        :  25.8235 ns per block ( 98.99%) -   0.4036 ns per byte -   4.1542 ns per structural -   2.4780 GB/s
../jsonexamples/canada.json
|    Speed        :  71.2842 ns per block ( 99.48%) -   1.1138 ns per byte -   7.4985 ns per structural -   0.8978 GB/s
|    Speed        :  70.6490 ns per block ( 99.64%) -   1.1039 ns per byte -   7.4316 ns per structural -   0.9059 GB/s
../jsonexamples/citm_catalog.json
|    Speed        :  22.9087 ns per block ( 99.57%) -   0.3580 ns per byte -   4.5464 ns per structural -   2.7936 GB/s
|    Speed        :  23.0030 ns per block ( 99.37%) -   0.3594 ns per byte -   4.5651 ns per structural -   2.7822 GB/s
../jsonexamples/github_events.json
|    Speed        :  24.5658 ns per block ( 98.16%) -   0.3840 ns per byte -   5.3711 ns per structural -   2.6044 GB/s
|    Speed        :  27.7358 ns per block ( 98.69%) -   0.4335 ns per byte -   6.0642 ns per structural -   2.3068 GB/s
../jsonexamples/google_maps_api_compact_response.json
|    Speed        :  77.7568 ns per block ( 96.82%) -   1.2178 ns per byte -   4.6150 ns per structural -   0.8211 GB/s
|    Speed        :  76.5027 ns per block ( 98.50%) -   1.1982 ns per byte -   4.5406 ns per structural -   0.8346 GB/s
../jsonexamples/google_maps_api_response.json
|    Speed        :  42.5392 ns per block ( 98.25%) -   0.6649 ns per byte -   5.5682 ns per structural -   1.5039 GB/s
|    Speed        :  41.9118 ns per block ( 98.72%) -   0.6551 ns per byte -   5.4860 ns per structural -   1.5264 GB/s
../jsonexamples/gsoc-2018.json
|    Speed        :  19.2111 ns per block ( 98.84%) -   0.3002 ns per byte -  13.1715 ns per structural -   3.3314 GB/s
|    Speed        :  19.3856 ns per block ( 99.10%) -   0.3029 ns per byte -  13.2911 ns per structural -   3.3014 GB/s
../jsonexamples/instruments.json
|    Speed        :  29.2367 ns per block ( 99.36%) -   0.4568 ns per byte -   3.7045 ns per structural -   2.1890 GB/s
|    Speed        :  29.0363 ns per block ( 99.16%) -   0.4537 ns per byte -   3.6791 ns per structural -   2.2041 GB/s
../jsonexamples/marine_ik.json
|    Speed        :  75.2808 ns per block ( 99.24%) -   1.1763 ns per byte -   5.4577 ns per structural -   0.8501 GB/s
|    Speed        :  76.1925 ns per block ( 99.27%) -   1.1905 ns per byte -   5.5238 ns per structural -   0.8400 GB/s
../jsonexamples/mesh.json
|    Speed        :  74.0608 ns per block ( 98.40%) -   1.1573 ns per byte -   5.4635 ns per structural -   0.8641 GB/s
|    Speed        :  76.7349 ns per block ( 98.40%) -   1.1991 ns per byte -   5.6607 ns per structural -   0.8340 GB/s
../jsonexamples/mesh.pretty.json
|    Speed        :  44.4144 ns per block ( 98.55%) -   0.6940 ns per byte -   7.1420 ns per structural -   1.4409 GB/s
|    Speed        :  45.6643 ns per block ( 98.96%) -   0.7135 ns per byte -   7.3430 ns per structural -   1.4015 GB/s
../jsonexamples/numbers100.json
|    Speed        :  67.7806 ns per block ( 99.89%) -   1.0591 ns per byte -   7.9481 ns per structural -   0.9442 GB/s
|    Speed        :  69.6803 ns per block ( 99.20%) -   1.0888 ns per byte -   8.1709 ns per structural -   0.9185 GB/s
../jsonexamples/numbers10.json
|    Speed        :  65.4156 ns per block ( 99.20%) -   1.0222 ns per byte -   7.6710 ns per structural -   0.9783 GB/s
|    Speed        :  66.7233 ns per block ( 98.47%) -   1.0426 ns per byte -   7.8244 ns per structural -   0.9592 GB/s
../jsonexamples/numbers.json
|    Speed        :  65.3789 ns per block ( 98.72%) -   1.0217 ns per byte -   7.6678 ns per structural -   0.9788 GB/s
|    Speed        :  64.7263 ns per block ( 97.26%) -   1.0115 ns per byte -   7.5913 ns per structural -   0.9886 GB/s
../jsonexamples/random.json
|    Speed        :  41.9050 ns per block ( 99.64%) -   0.6548 ns per byte -   3.7979 ns per structural -   1.5271 GB/s
|    Speed        :  40.9663 ns per block ( 99.64%) -   0.6402 ns per byte -   3.7128 ns per structural -   1.5621 GB/s
../jsonexamples/repeat.json
|    Speed        :  49.7079 ns per block ( 98.39%) -   0.7791 ns per byte -   8.7001 ns per structural -   1.2835 GB/s
|    Speed        :  49.0899 ns per block ( 98.48%) -   0.7695 ns per byte -   8.5919 ns per structural -   1.2996 GB/s
../jsonexamples/tree-pretty.json
|    Speed        :  38.8685 ns per block ( 98.21%) -   0.6074 ns per byte -   5.3016 ns per structural -   1.6464 GB/s
|    Speed        :  38.2056 ns per block ( 98.55%) -   0.5970 ns per byte -   5.2112 ns per structural -   1.6750 GB/s
../jsonexamples/twitter10.json
|    Speed        :  28.8527 ns per block ( 99.38%) -   0.4508 ns per byte -   5.1517 ns per structural -   2.2181 GB/s
|    Speed        :  29.0336 ns per block ( 99.61%) -   0.4537 ns per byte -   5.1840 ns per structural -   2.2043 GB/s
../jsonexamples/twitter_api_compact_response.json
|    Speed        :  55.6962 ns per block ( 98.13%) -   0.8734 ns per byte -   7.1661 ns per structural -   1.1449 GB/s
|    Speed        :  55.7152 ns per block ( 98.19%) -   0.8737 ns per byte -   7.1686 ns per structural -   1.1445 GB/s
../jsonexamples/twitter_api_response.json
|    Speed        :  45.2594 ns per block ( 98.05%) -   0.7092 ns per byte -   7.5170 ns per structural -   1.4101 GB/s
|    Speed        :  45.0669 ns per block ( 96.12%) -   0.7062 ns per byte -   7.4851 ns per structural -   1.4161 GB/s
../jsonexamples/twitterescaped.json
|    Speed        :  56.4796 ns per block ( 99.61%) -   0.8825 ns per byte -   8.9815 ns per structural -   1.1331 GB/s
|    Speed        :  56.1566 ns per block ( 99.58%) -   0.8775 ns per byte -   8.9301 ns per structural -   1.1396 GB/s
../jsonexamples/twitter.json
|    Speed        :  26.3584 ns per block ( 99.05%) -   0.4119 ns per byte -   4.7067 ns per structural -   2.4279 GB/s
|    Speed        :  26.3824 ns per block ( 99.15%) -   0.4122 ns per byte -   4.7110 ns per structural -   2.4257 GB/s
../jsonexamples/twitter_timeline.json
|    Speed        :  46.5348 ns per block ( 98.60%) -   0.7272 ns per byte -   5.7655 ns per structural -   1.3751 GB/s
|    Speed        :  46.9106 ns per block ( 98.33%) -   0.7331 ns per byte -   5.8121 ns per structural -   1.3641 GB/s
../jsonexamples/update-center.json
|    Speed        :  33.8479 ns per block ( 99.00%) -   0.5289 ns per byte -   4.4464 ns per structural -   1.8908 GB/s
|    Speed        :  32.9145 ns per block ( 99.35%) -   0.5143 ns per byte -   4.3238 ns per structural -   1.9444 GB/s
```


Looking at the two trials, we have that this PR is marginally slower on instruments, random.json, tree-pretty.json, update-center.json,  while it is faster on mesh.pretty.json, mesh, numbers100.json, numbers10.json... On other files, it is mostly inconclusive, it depends on how often you run the trial.


Now we can look at instructions instead...

```
$ for i in ../jsonexamples/*.json ; do echo $i; ./benchmark/parse -n 100 $i | grep Instru | head -n 1; ./../buildmaster/benchmark/parse -n 100 $i | grep Instr | head -n 1; done
../jsonexamples/apache_builds.json
|    Instructions : 329.8361 per block    ( 99.99%) -   5.1545 per byte    -  53.0608 per structural    -    3.389 per cycle
|    Instructions : 327.6365 per block    ( 99.99%) -   5.1202 per byte    -  52.7070 per structural    -    3.431 per cycle
../jsonexamples/canada.json
|    Instructions : 902.6156 per block    (100.00%) -  14.1035 per byte    -  94.9470 per structural    -    3.454 per cycle
|    Instructions : 902.6152 per block    (100.00%) -  14.1035 per byte    -  94.9469 per structural    -    3.437 per cycle
../jsonexamples/citm_catalog.json
|    Instructions : 305.2628 per block    (100.00%) -   4.7698 per byte    -  60.5812 per structural    -    3.594 per cycle
|    Instructions : 304.3761 per block    (100.00%) -   4.7560 per byte    -  60.4052 per structural    -    3.595 per cycle
../jsonexamples/github_events.json
|    Instructions : 284.8350 per block    ( 99.98%) -   4.4519 per byte    -  62.2771 per structural    -    3.132 per cycle
|    Instructions : 283.9931 per block    ( 99.98%) -   4.4388 per byte    -  62.0930 per structural    -    3.157 per cycle
../jsonexamples/google_maps_api_compact_response.json
|    Instructions : 720.4108 per block    ( 99.95%) -  11.2831 per byte    -  42.7578 per structural    -    3.020 per cycle
|    Instructions : 715.7676 per block    ( 99.95%) -  11.2104 per byte    -  42.4822 per structural    -    3.104 per cycle
../jsonexamples/google_maps_api_response.json
|    Instructions : 381.8431 per block    ( 99.96%) -   5.9686 per byte    -  49.9814 per structural    -    3.037 per cycle
|    Instructions : 379.7377 per block    ( 99.96%) -   5.9357 per byte    -  49.7058 per structural    -    3.054 per cycle
../jsonexamples/gsoc-2018.json
|    Instructions : 180.8914 per block    (100.00%) -   2.8265 per byte    - 124.0225 per structural    -    2.541 per cycle
|    Instructions : 180.7656 per block    (100.00%) -   2.8245 per byte    - 123.9363 per structural    -    2.519 per cycle
../jsonexamples/instruments.json
|    Instructions : 367.5850 per block    ( 99.99%) -   5.7437 per byte    -  46.5755 per structural    -    3.411 per cycle
|    Instructions : 366.5899 per block    ( 99.99%) -   5.7281 per byte    -  46.4494 per structural    -    3.422 per cycle
../jsonexamples/marine_ik.json
|    Instructions : 827.5161 per block    (100.00%) -  12.9300 per byte    -  59.9932 per structural    -    2.983 per cycle
|    Instructions : 828.0154 per block    (100.00%) -  12.9378 per byte    -  60.0294 per structural    -    2.945 per cycle
../jsonexamples/mesh.json
|    Instructions : 880.6041 per block    (100.00%) -  13.7604 per byte    -  64.9620 per structural    -    3.223 per cycle
|    Instructions : 880.6040 per block    (100.00%) -  13.7604 per byte    -  64.9620 per structural    -    3.107 per cycle
../jsonexamples/mesh.pretty.json
|    Instructions : 533.7253 per block    (100.00%) -   8.3397 per byte    -  85.8249 per structural    -    3.259 per cycle
|    Instructions : 533.7252 per block    (100.00%) -   8.3397 per byte    -  85.8249 per structural    -    3.165 per cycle
../jsonexamples/numbers100.json
|    Instructions : 732.1301 per block    (100.00%) -  11.4395 per byte    -  85.8514 per structural    -    2.876 per cycle
|    Instructions : 732.1301 per block    (100.00%) -  11.4395 per byte    -  85.8514 per structural    -    2.855 per cycle
../jsonexamples/numbers10.json
|    Instructions : 732.2108 per block    (100.00%) -  11.4412 per byte    -  85.8634 per structural    -    3.036 per cycle
|    Instructions : 732.2104 per block    (100.00%) -  11.4412 per byte    -  85.8634 per structural    -    2.971 per cycle
../jsonexamples/numbers.json
|    Instructions : 733.6066 per block    (100.00%) -  11.4641 per byte    -  86.0391 per structural    -    3.036 per cycle
|    Instructions : 733.6027 per block    (100.00%) -  11.4641 per byte    -  86.0387 per structural    -    3.064 per cycle
../jsonexamples/random.json
|    Instructions : 524.0109 per block    (100.00%) -   8.1885 per byte    -  47.4912 per structural    -    3.389 per cycle
|    Instructions : 521.0006 per block    (100.00%) -   8.1415 per byte    -  47.2184 per structural    -    3.438 per cycle
…