@jkeiser 2020-05-03
ASCII encoded bits, that is devilishly clever.