@lemire 2020-05-03
> ASCII encoded bits, that is devilishly clever.

All of our emails get base64 encoded... but, base64 encoding would be overkill here. Still, if we wanted to write lots of data on the string tape and still have it be valid utf8... we could use base64.

I went looking for something more specialized and I wrote this...
 https://lemire.me/blog/2020/05/02/encoding-binary-in-ascii-very-fast/