@lemire 2020-05-07
@jkeiser  This PR is now probably correct, but there is a real performance regression in stage 2. It is kind of obvious where it is, since we only change a few lines in stage 2. Figuring out how to restore the performance is going to be great fun.

In theory it should be faster since our main path is simpler.  But I have had to add a damn branch.