@lemire 2020-05-07
Still a work in progress but getting closer.