@jkeiser 2020-05-07
Out of curiosity, did either or both of those things help? I'm particularly curious if "make an unlikely one-legged branch" did anything ...