@lemire 2020-05-07
> Out of curiosity, did either or both of those things help? I'm particularly curious if "make an unlikely one-legged branch" did anything ...

The price of an extra branch is hard to hide entirely. I am still struggling. 

However, I think we want to make absolutely certain that the "never taken" path is not inlined.

My last commit absolutely helped. I think I have greatly reduced the performance degradation... to something tiny. I think our perf. tests are failing because they are very sensitive.

That's fine. I will work on it again.