@lemire 2020-05-07
@jkeiser  This is ready for review. I think that the performance failures are just noise at this point.