@jkeiser 2020-05-07
@lemire on my machine, I see a consistent 5% performance loss.

Master:

```
All Stages
|    Speed        :  23.2671 ns per block ( 83.37%) -   0.3636 ns per byte -   4.1547 ns per structural -   2.7505 GB/s
|- Stage 1
|    Speed        :  11.1471 ns per block ( 39.94%) -   0.1742 ns per byte -   1.9905 ns per structural -   5.7410 GB/s
|- Stage 2
|    Speed        :  12.0896 ns per block ( 43.32%) -   0.1889 ns per byte -   2.1588 ns per structural -   5.2935 GB/s

4355.4 documents parsed per second (best)
```

This branch:

```
All Stages
|    Speed        :  24.4224 ns per block ( 88.10%) -   0.3816 ns per byte -   4.3610 ns per structural -   2.6204 GB/s
|- Stage 1
|    Speed        :  11.4410 ns per block ( 41.27%) -   0.1788 ns per byte -   2.0430 ns per structural -   5.5936 GB/s
|- Stage 2
|    Speed        :  12.9206 ns per block ( 46.61%) -   0.2019 ns per byte -   2.3071 ns per structural -   4.9531 GB/s

4149.4 documents parsed per second (best)
```

Let me check skylake/skylake-x for good measure.