@jkeiser 2020-05-08
Checking perf on skylake now that you learned me some Linux :)

I wonder if this can be done branchlessly if we use a separate type for long strings ...

1. Append string to buffer.
2. Append \0 to buffer.
3. Append high byte of length (len >> 24) to buffer.
4. [Trick] If high byte is nonzero:
    - Set the buffer after the high byte instead of the \0: `buffer += ((len >> 24) > 0)`.
    - Set the type to # (" + 1, e.g. `type = '"' + ((len >> 24) > 0)`).
5. Set tape = the type (" or #) + low 3 bytes of length + string index.

The reader has to do an unlikely branch still :/ But they had to before anyway.