@lemire 2020-05-08
It seems that you are writing part of the string length at the end of the string (Append high byte of length (len >> 24) to buffer.). But how does that help you?

I give you the string, you need to figure out its length... you only have 3 bytes out of the 4 bytes of the length... You have written the missing byte of information at a position that is "length" away.