@jkeiser 2020-05-08
The key is the type. If the type of element is ", then the other three bytes are the length. If it is #, then you need to read the fourth byte from the buffer