@lemire 2020-05-08
I mean, yes, the fourth byte is on the string buffer, somewhere... but how do you know which byte it is?