@lemire 2020-05-08
Oh. I think I understand what you mean now. You want us to point at the end of the string and then backtrack.