@lemire 2020-05-08
It could be used to avoid the nasty thing I did with long strings...