@lemire 2020-05-08
@jkeiser I have updated the PR with your strategy. It saves me from having to move the string data. It may cost extra branching at query time.