@lemire 2020-05-09
@jkeiser On GNU GCC 8.3, this code is a wash performance wise. Certainly not a regression.

I have re-enabled checkperf in circleci with other compilers and it looks like some of them pass. So this would be a regression on older compilers. Fine.

Now, we do not proceed with this or something similar, we will not be able to lift UTF8 from stage 1. Admittedly, that's only going to benefit some files and maybe not that much... So maybe this whole PR is better dropped...

But that's the thing. If we don't do this, then we better drop the issue about moving UTF8 validation off stage 1 because I don't know how else to do it.

Note: dropping this PR would be fine.