@lemire 2020-05-09
I am starting to think that this should be dropped and that https://github.com/simdjson/simdjson/issues/185 should be closed.