@lemire 2020-05-09
I am going to close this. Gains are probably too minor.