@jkeiser 2020-05-09
Are you not seeing the regression I saw at all? I'm ok with a regression on older compilers personally. But I was running with gcc 9.3.0, I believe.

Here's the thing I've been thinking about. I don't know how we make progress towards (possibly) bigger improvements like stage 2 utf8 validation if a necessary incremental step like this reduces performance. I feel like we might be throwing out the baby before we even put it in the bathwater, and I feel bad being a part of possibly stifling it.

Not that I mind dropping it if we don't see any benefit now *or* later ... but I don't think we know or suspect that yet, do we?