@jkeiser 2020-05-09
@lemire on skylake and skylake-x GCC 8.4.0, I see zero performance difference with this patch on twitter.json. (This is with [the version I tested before](https://github.com/simdjson/simdjson/pull/856/commits/6c0525bda69d7a8d46f39a7962f821a3fefc9ab4), not the one where you tried to implement my half-baked algorithm :)

It doesn't seem like a GCC 9 thing either; on GCC 9.3.0 the difference is within the level of noise.