@jkeiser 2020-05-09
I think we should move forward with this or something like it, given the above information. Using less tape is definitely desirable. No performance loss means somebody somewhere is going to win (probably users).